+++ annotations/src/main/java/org/realityforge/arez/annotations/OnDepsUpdated.java
@@ -6,10 +6,7 @@ import java.lang.annotation.Target;
 import javax.annotation.Nonnull;
 
 /**
- * Identifies method that will be invoked when the dependencies of the paired Tracked are updated.
- *
- * i.e. If the Tracked method is named "render" then the Reaction will default
- * to being named "onRenderDepsUpdate".
+ * Identifies method that will be invoked when the dependencies of the paired @Track annotated method are updated.
  *
  * <p>The method that is annotated with OnDepsUpdated must comply with the additional constraints:</p>
  * <ul>
@@ -29,9 +26,10 @@ public @interface OnDepsUpdated
   /**
    * Return the name of the paired Tracked relative to the component.
    * The value must conform to the requirements of a java identifier.
-   * The name need not be specified .
+   * The name need not be specified. If the @Track annotated method is
+   * named "render" then this will default to being named "onRenderDepsUpdate".
    *
-   * @return the name of the paired Tracked relative to the component.
+   * @return the name of the paired @Track annotated method relative to the component.
    */
   @Nonnull
   String name() default "<default>";