+++ processor/src/test/resources/bad_input/com/example/on_deps_updated/OnDepsUpdatedNotStaticModel.java
@@ -2,12 +2,12 @@ package com.example.on_deps_updated;
 
 import org.realityforge.arez.annotations.ArezComponent;
 import org.realityforge.arez.annotations.OnDepsUpdated;
-import org.realityforge.arez.annotations.Tracked;
+import org.realityforge.arez.annotations.Track;
 
 @ArezComponent
 public class OnDepsUpdatedNotStaticModel
 {
-  @Tracked
+  @Track
   public void render()
   {
   }