+++ processor/src/test/resources/bad_input/com/example/tracked/TrackedBadNameModel.java
@@ -2,12 +2,12 @@ package com.example.tracked;
 
 import org.realityforge.arez.annotations.ArezComponent;
 import org.realityforge.arez.annotations.OnDepsUpdated;
-import org.realityforge.arez.annotations.Tracked;
+import org.realityforge.arez.annotations.Track;
 
 @ArezComponent
 public class TrackedBadNameModel
 {
-  @Tracked(name = "-ace")
+  @Track(name = "-ace")
   static void render()
   {
   }