+++ processor/src/test/resources/bad_input/com/example/tracked/TrackedDuplicatedModel.java
@@ -2,17 +2,17 @@ package com.example.tracked;
 
 import org.realityforge.arez.annotations.ArezComponent;
 import org.realityforge.arez.annotations.OnDepsUpdated;
-import org.realityforge.arez.annotations.Tracked;
+import org.realityforge.arez.annotations.Track;
 
 @ArezComponent
 public class TrackedDuplicatedModel
 {
-  @Tracked
+  @Track
   public void render()
   {
   }
 
-  @Tracked( name = "render" )
+  @Track( name = "render" )
   public void render2()
   {
   }