+++ processor/src/test/resources/bad_input/com/example/tracked/TrackedMissingOnDepsUpdated.java
@@ -1,12 +1,12 @@
 package com.example.tracked;
 
 import org.realityforge.arez.annotations.ArezComponent;
-import org.realityforge.arez.annotations.Tracked;
+import org.realityforge.arez.annotations.Track;
 
 @ArezComponent
 public class TrackedMissingOnDepsUpdated
 {
-  @Tracked
+  @Track
   void render()
   {
   }