+++ processor/src/test/resources/input/com/example/tracked/BasicTrackedWithExceptionsModel.java
@@ -3,12 +3,12 @@ package com.example.tracked;
 import java.text.ParseException;
 import org.realityforge.arez.annotations.ArezComponent;
 import org.realityforge.arez.annotations.OnDepsUpdated;
-import org.realityforge.arez.annotations.Tracked;
+import org.realityforge.arez.annotations.Track;
 
 @ArezComponent
 public class BasicTrackedWithExceptionsModel
 {
-  @Tracked
+  @Track
   public void render()
     throws ParseException
   {